requirements/common.txt
fabric==1.8.3
pyyaml==3.10
Markdown==2.3.1
Pygments==1.5
flup==1.0.2
django-extjs4==1.1.4-extjs4.1.1
gunicorn==0.14.6
django-seleniumhelpers==1.1.1
django-crispy-forms==1.4.0
openpyxl==1.6.1
django==1.6.5
URLObject==2.0.1
django-simple-rest==1.4.1
mimeparse==0.1.3
django_errortemplates==1.0.1
numpy==1.6.2
whoosh==2.5.7
amqp==1.3.3
anyjson==0.3.3
billiard==3.3.0.5
celery==3.1.1
django-celery==3.1.1
kombu==3.0.4
python-dateutil==1.5
pytz==2012j
celery-haystack==0.7.2
django-appconf==0.6
django-celery-transactions==0.1.3
six==1.2.0
django-haystack==2.1.0
httplib2==0.7.7
South==0.8.2
django-celery-email==1.0.4
mock==1.0.1
django_cradmin==1.0.0-alpha.010


# Local modules
-e ../thirdpartylibs/djangorestframework
-e .
